Move blazor only functions to Boot.WebAssembly.ts
